:sparkles: add sizeSX in avatar
