+++ packages/react/src/components/Avatar/avatar.tsx
@@ -1,12 +1,13 @@
 import { useTheme } from 'styled-components';
 
 import Text from '@components/Text';
+import { SizeSX } from '@styles/size';
 import { SpacingSX } from '@styles/spacing';
 import { Palette } from '@styles/theme';
 
 type AvatarSize = 'small' | 'medium' | 'large';
 
-interface AvatarSX extends SpacingSX {
+interface AvatarSX extends SpacingSX, SizeSX {
   border?: string;
   borderRadius?: string;
   backgroundColor?: keyof Palette;
@@ -68,6 +69,21 @@ const getSizeStyle = (size: AvatarSize) => {
   }
 };
 
+/**
+ *
+ * @name Avatar
+ * @description `src` 속성에 이미지를 넣는 경우 `size` 에 맞는 크기로 동그란 컴포넌트가 생성됩니다.
+만약 `src` 가 지정되지 않고, `children` 으로 문자열 을 사용하는 경우 글자수에 맞게 문자열 아바타가 생성됩니다.
+ * @props `src` : `Avatar` 에 적용할 이미지의 경로를 입력합니다.
+ * @props `alt` : 이미지의 텍스트 설명입니다.
+ * @props `size` : `small`, `medium`, `large` 중 선택할 수 있습니다. `default: medium`
+ * @props `sx` : `border`, `borderRadius`, `backgroundColor`, `color` 이외 `Spacing`, `Size` 관련 속성을 활용할 수 있습니다.
+ * @example
+ * ```tsx
+ *  <Avatar src={image.src} alt={image.name} size="large" />
+ *  <Avatar sx={{ backgroundColor: 'primary' }}>박상진</Avatar>
+ * ```
+ */
 const Avatar = (props: AvatarProps) => {
   const theme = useTheme();
   const { alt, src, sx = {}, size = 'medium', children, className } = props;
