+++ packages/react/src/styles/size.ts
@@ -1,0 +1,8 @@
+export type SizeSX = {
+  width?: string;
+  maxWidth?: string;
+  minWidth?: string;
+  height?: string;
+  maxHeight?: string;
+  minHeight?: string;
+};
